Crate egui_wgpu

This crates provides bindings between egui and wgpu.

Feature flags

Re-exports

pub use renderer::Renderer;

pub use renderer::Callback;

pub use renderer::CallbackResources;

pub use renderer::CallbackTrait;

pub use wgpu;

Modules

renderer

Low-level painting of egui on wgpu.

Structs

RenderState

Access to the render state for egui.

WgpuConfiguration

Configuration for using wgpu with eframe or the egui-wgpu winit feature.

Enums

SurfaceErrorAction

Specifies which action should be taken as consequence of a wgpu::SurfaceError

WgpuError

Functions

depth_format_from_bits

Take’s epi’s depth/stencil bits and returns the corresponding wgpu format.

preferred_framebuffer_format

Find the framebuffer format that egui prefers